requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.30.0
cffi==2.0.0
click==8.2.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.116.2
greenlet==3.2.4
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
pwdlib==0.2.1
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
python-slugify==8.0.4
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
